move version_data.py to match other services
